перенос layout`ов в отдельный модуль /сам модуль требует переимпорта gui и images, т.к. ссылается на них
